requirements.txt
Django==4.2.7
django-mptt==0.15.0
